interpreter/requirements.txt
pinecone-client
langchain-pinecone
langchain-huggingface
sentence-transformers
langchain_text_splitters